--- src/constant/authApiError.ts
@@ -1,6 +1,0 @@
-export const AUTH_API_ERROR = {
-  EXISTS: 'exists',
-  NO_DATA: 'no-data',
-  SERVER_ERROR: 'server-error',
-  INVALID: 'invalid',
-} as const;
