+++ src/constants/index.ts
@@ -1,0 +1,1 @@
+export { AUTH_API_ERROR, STATUS_COLORS } from './constants';
