+++ src/entities/schemas/index.ts
@@ -1,0 +1,2 @@
+export { Client } from './client';
+export { getAuthSchema } from './auth';
